move DurationConverter to separate module
